@@ -12,10 +12,11 @@
   "author": "Darin Haener <dphaener@gmail.com> (https://github.com/dphaener)",
   "license": "MIT",
   "dependencies": {
-    "alt": "^0.17.3",
     "immutable": "^3.7.5",
-    "react": ">=0.13",
-    "react-dnd": "^1.1.5",
+    "react": "^0.14.6",
+    "react-dom": "^0.14.6",
+    "react-draggable": "^1.1.3",
+    "react-grid-layout": "0.10.0-beta1",
     "strftime": "^0.9.2"
   },
   "devDependencies": {
@@ -31,10 +32,10 @@
     "file-loader": "0.8.4",
     "img-loader": "1.1.0",
     "mocha": "2.2.5",
-    "mocha-babel": "3.0.0",
+    "mocha-babel": "3.0.3",
     "node-jsdom": "3.1.5",
     "node-libs-browser": "0.5.2",
-    "node-sass": "3.2.0",
+    "node-sass": "^3.4.2",
     "react-hot-loader": "1.2.7",
     "react-tools": "0.13.3",
     "sass-loader": "1.0.2",
